refactor: Further reorganize ensure files
